+++ lib/src/pluto_grid_configuration.dart
@@ -9,6 +9,11 @@ class PlutoGridConfiguration {
   /// border between columns.
   final bool enableColumnBorder;
 
+  /// Activate the shadow that separates each area of the grid.
+  final bool enableGridBorderShadow;
+
+  final BorderRadiusGeometry gridBorderRadius;
+
   final Color gridBackgroundColor;
 
   /// Grid border color. (Grid outline color, Frozen column division line color)
@@ -72,6 +77,8 @@ class PlutoGridConfiguration {
 
   PlutoGridConfiguration({
     this.enableColumnBorder = false,
+    this.enableGridBorderShadow = true,
+    this.gridBorderRadius = BorderRadius.zero,
     this.gridBackgroundColor = Colors.white,
     this.gridBorderColor = const Color(0xFFA1A5AE),
     this.activatedColor = const Color(0xFFDCF5FF),
@@ -106,6 +113,8 @@ class PlutoGridConfiguration {
 
   PlutoGridConfiguration.dark({
     this.enableColumnBorder = false,
+    this.enableGridBorderShadow = true,
+    this.gridBorderRadius = BorderRadius.zero,
     this.gridBackgroundColor = const Color(0xFF111111),
     this.gridBorderColor = const Color(0xFF000000),
     this.activatedColor = const Color(0xFF313131),